Make JRE Source Optional
Previously, that JRE source was a required parameter to the application.  A
recent problem where the JRE source could not be found in its usual place
(because it was a server JRE) indicated that staging should no fail completely
if a JRE source can't be found.  This change makes that parameter optional so
that staging can progress even in the face of a missing JRE source.

[cloudfoundry/java-buildpack#415]